linux/ena: Count number of unique tx packets missing completion
Currently, we're reporting the number of missing tx packets including
duplicates in check_missing_comp_in_tx_queue() since we increment the
missing_tx counter before we check if this packet has been encountered
in previous calls.

Modify the counter to be incremented only in case this packet is not
seen before (tx_buf->print_once is not set).

Fixes: 6733e9e9cea8 ("linux/ena: Add more information on TX timeouts")
Signed-off-by: Osama Abboud <osamaabb@amazon.com>