linux/ena: Revert napi_consume_skb() and napi_build_skb() usage
napi_consume_skb() serves two purposes:

1. Store freed/consumed SKBs in a CPU-specific cache and
   free them as a bulk.
2. From the introduction of napi_build_skb(), the SKBs stored
   in the cache are to be re-utilized on the RX path
   (when SKBs are allocated in order to populate the incoming data
   into them) in order to reduce the amount of allocations on
   this path.

During testing, PPS degradation has been identified when running
on instance types such as p4d.
The degradation was root-caused to the introduction of
napi_consume_skb() in the ENA driver.

In order to avoid the degradation, the API usage has been rolled
back (The change was introduced in commit [1]).

In addition, as napi_build_skb() utilizes the above cache, and is
tightly coupled with napi_consume_skb(), its usage has been
rolled back as well. (The change was introduced in commit [2]).

[1]: commit fe54212d5c31 ("linux/ena: Add napi_consume_skb")
[2]: commit 9175a55483ea ("linux/ena: Add napi_build_skb support")

Signed-off-by: David Arinzon <darinzon@amazon.com>