linux/ena: Add SUSE kernel subversion parsing
SUSE kernels are of the format A.B.C-D.E.F (F is optional
depending on the release. There's also .G, but it's not used
to differentiate between versions). It is currently not possible to
differentiate between the versions as the current macros only allow
access to A.B.C.
Updating the Makefile to add SUSE_KERNEL_SUBVERSION_X (X is an
index starting from 1) to the compilation flags in order to be
used in kcompat.h.

Signed-off-by: David Arinzon <darinzon@amazon.com>