linux/ena: Fix incorrect dma sync when SWIOTLB is on
Background:
===========
When SWIOTLB is on and a memory area is mapped for DMA, there are 2
memory areas involved:
1. Memory area used by the device.
2. Memory area used by the CPU.

The SWIOTLB synchronizes between the 2 memory areas when dma_sync_*
functions are called, by copying data from one memory area to the other.
For example, when we call dma_sync_single_for_cpu(), SWIOTLB copies from
the device memory area to the CPU memory area.

dma_sync_single_for_cpu(), can be called for part of the DMA-mapped
memory (say we can sync 128 bytes out of a 4K DMA-mapped page).

dma_unmap_page() also calls dma_sync_single_for_cpu(), triggering
the copy of the entire DMA-mapped page from the device memory area
to the CPU memory area.

When we use DRB, we reuse a single DMA-mapped RX page for multiple RX
buffers.

Issue Description:
==================

A. Flow for a single RX buffer when the RX page can be
   reused for more RX buffers
-------------------------------------------------------
Below is a diagram of a DMA-mapped 4K page, used for multiple RX buffers,
when SWIOTLB is turned on.
Upper part is the page used by the CPU and lower part is the page used by
the device.
A single RX buffer with all of its parts is enclosed with
"||" from both sides.
At time t=0, we are using the middle of the page as an RX buffer after
previously using the lower parts of the page for RX buffers. In this flow
there is still enough space in the page for it to be reused again for
RX buffers.

This flow's steps are (steps (1)-(3) are marked in the diagram):
(1) The device writes to the data part of the current buffer in the
    page area used by the device.
(2) dma_sync_single_for_cpu() is called to copy the data from the
    device page to the CPU page.
(3) build_skb() creates the skb around the data of the first buffer
    of the packet. It writes the skb_shared_info part of the skb, which
    includes pointers to the fragments of the packet, in the tailroom.
    This tailroom is not copied to the device side of the memory area
    because it is unnecessary for the device, so the tailroom on the
    device side is full of uninitialized memory.

                                                          (3)build_skb()
CPU side memory area                                              ||
                                                                  \/
 0                                                                                    4K
 --------------------------------------------------------------------------------------
 | < previous || rx_ring->   |    pkt_offset    |   < data >   | tailroom        ||...|
 |  buffers > || rx_headroom |                  |              | skb_shared_info ||   |
 --------------------------------------------------------------------------------------

                                                      /\
                                                      ||
                                        (2)dma_sync_single_for_cpu()
                                                      ||
 0                                                                                     4K
 ---------------------------------------------------------------------------------------
 | < previous || rx_ring->   |    pkt_offset    |   < data >   | tailroom         ||...|
 |  buffers > || rx_headroom |                  |              | GARBAGE          ||   |
 ---------------------------------------------------------------------------------------

Device side memory area                                /\
                                                       ||
                                        (1) device writes RX packet data

B. Flow for a single RX buffer when the RX page can
   not be reused for more RX buffers
---------------------------------------------------
For this flow, we zoom out of the previous diagram. Each buffer used is
divided into two parts: the beginning, and the tailroom. The CPU side page
has tailrooms filled with skb_shared_info and the device-side tailrooms
are filled with uninitialized memory.

At t=0, we are using up the last possible buffer in this buffer page.
For this case, step (2) from before is replaced with dma_page_unmap().
dma_page_unmap() calls dma_sync_single_for_cpu() for the ENTIRE PAGE,
including copying the GARBAGE tailroom for all previous RX packets
received using this page.
If any of the skbs pointing to buf_a or buf_b have not yet been handled
by the kernel, their skb_shared_info will be overridden with garbage by
this call to dma_page_unmap(). When the kernel finally handles these
skbs, it will try to access garbage pointers and crash.

                                                                     (3)build_skb()
CPU side memory area                                                       ||
                                                                           \/
 0                                                                                                4K
 --------------------------------------------------------------------------------------------------
 | buf_a     | buf_a           || buf_b     | buf_b           || buf_c     | buf_c           ||...|
 | beginning | skb_shared_info || beginning | skb_shared_info || beginning | skb_shared_info ||   |
 --------------------------------------------------------------------------------------------------

                                                      /\
                                                      ||
                                              (2)dma_unmap_page()
                                                      ||
 0                                                                                                4K
 --------------------------------------------------------------------------------------------------
 | buf_a     | buf_a           || buf_b     | buf_b           || buf_c     | buf_c           ||...|
 | beginning | GARBAGE         || beginning | GARBAGE         || beginning | GARBAGE         ||   |
 --------------------------------------------------------------------------------------------------

Device side memory area                                             /\
                                                                    ||
                                                      (1) device writes RX packet data

Solution:
=========
Do not sync the whole page when calling dma_unmap_page() by replacing it
with dma_unmap_page_attrs(), and passing it the DMA_ATTR_SKIP_CPU_SYNC
attribute, which tells it not to sync the page before unmapping.

Call dma_sync_single_for_cpu() explicitly to sync the data of the last RX
buffer of the page.

Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>
Change-Id: I8f8c4a84abac92f52af39609c4b55db730efdb6b