linux/ena: Fix RHEL7 compilation issues due to missing dma_unmap_page_attrs()
Background
==========
Commit [1] replaced the older representation of dma attributes
from struct dma_attr to an unsigned long of flags.

Commit [2] added dma_unmap_page_attr(), which receives dma attributes.

Both of these commits are needed for the code to compile
without any changes.

RHEL7 issues and solutions
==========================

Issue 1
-------
RHEL7 versions <= 7.9 have the function dma_unmap_page_attr()
(which was added in commit[2]) but do not have commit [1].
We solve this issue in 2 parts:

Part 1
``````
For these kernels, this commit defines the wrapper function
ena_dma_unmap_page_attrs(), whose purpose is to transform the
attributes input from unsigned long attributes to struct dma_attrs
attributes, where needed.

Part 2
``````
Kernels that have older representation of dma attributes as
struct dma_attrs, also use enum dma_attrs which looks like this (actual
integer numbers of each enum value are in comments):

enum dma_attr {
        DMA_ATTR_WRITE_BARRIER,         /* 0 */
        DMA_ATTR_WEAK_ORDERING,         /* 1 */
        DMA_ATTR_WRITE_COMBINE,         /* 2 */
        DMA_ATTR_NON_CONSISTENT,        /* 3 */
        DMA_ATTR_NO_KERNEL_MAPPING,     /* 4 */
        DMA_ATTR_SKIP_CPU_SYNC,         /* 5 */
        DMA_ATTR_FORCE_CONTIGUOUS,      /* 6 */
        DMA_ATTR_ALLOC_SINGLE_PAGES,    /* 7 */
        DMA_ATTR_MAX,                   /* 8 */
};

Kernels that have newer representation of dma attributes as unsigned long
have the following defines instead of the above enum:

        #define DMA_ATTR_WEAK_ORDERING          (1UL << 1)
        #define DMA_ATTR_WRITE_COMBINE          (1UL << 2)
        #define DMA_ATTR_NO_KERNEL_MAPPING      (1UL << 4)
        #define DMA_ATTR_SKIP_CPU_SYNC          (1UL << 5)
        #define DMA_ATTR_FORCE_CONTIGUOUS       (1UL << 6)
        #define DMA_ATTR_ALLOC_SINGLE_PAGES     (1UL << 7)
        #define DMA_ATTR_NO_WARN                (1UL << 8)
        #define DMA_ATTR_PRIVILEGED             (1UL << 9)

Our goal in this commit is to use the second notation and convert it
to the older one only for the older kernels in as few places as
possible.

To achieve this we want to be able to pass flags of the newer notation
to functions in the driver code, allowing multiple flags to be passed to
the functions, like so: (DMA_ATTR_WEAK_ORDERING | DMA_ATTR_NON_CONSISTENT)

In newer kernels (defines) this will be translated to 0001 | 1000 = 1001
In older kernels (enum) this will be translated to 0001 | 0011 = 0011

In older kernels if we use this kind of combination of flags we lose data
(we lost the flag DMA_ATTR_WEAK_ORDERING).

To solve this issue we added our own #defines in kcompat.h, which
have the same names as the current #define/enum values with "ENA_"
added in the beginning.
The values of these defines are the same as the values of the newer
kernel defines and we use them the same way in both old and new kernels.
The difference comes when we translate for older kernels them to struct
dma_attrs flags, in one place where it is necessary.

We use new names starting with "ENA_" for these defines, instead of
using the original names, to avoid collision between the names
of the defines and the same exact names in the enum values at
compilation time.

Issue 2
-------
RHEL versions <= 7.4 don't have both commits [1] and [2].
Therefore in addition to the changes added for RHEL <= 7.9 this
commit provides the implementation of dma_unmap_page_attr(),
copied as is from commit [2] with the attr parameter changed
from unsigned long to struct dma_attrs.

kernel-alt
----------
RHEL7 kernel can be replaced by installing the kernel-alt package,
which upgrades the kernel to 4.14.0-115 (see [3]).
This commit addresses compilation issues when this kernel is installed.

[1]: https://github.com/torvalds/linux/commit/00085f1efa387a8ce100e3734920f7639c80caa3
[2]: https://github.com/torvalds/linux/commit/0495c3d367944e4af053983ff3cdf256b567b053
[3]: https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/7/html-single/7.6_release_notes/index

Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>
Change-Id: I55d1bc0faf357ba76f00222f3f1b43d6ec6242a2