linux/ena: Fix missing dma attribute flag in RHEL 6
This commit handles one last issue that appears in RHEL6.
RHEL 6 does not have DMA_ATTR_SKIP_CPU_SYNC flag in enum
dma_attr. Therefore ENA_DMA_ATTR_SKIP_CPU_SYNC cannot
be a translation of DMA_ATTR_SKIP_CPU_SYNC.

Since RHEL6 kernel is 2.6 and is very old, it was decided
that we are not going to make the swiotlb functionality
work in RHEL6, by adding all the missing code.
Instead, this commit defines ENA_DMA_ATTR_SKIP_CPU_SYNC
to be 0, effectively making dma_unmap_page_attrs() with
this flag set, to be equivalent to dma_unmap_page().
This means RHEL6 is unusable with swiotlb on, but when it
is off, it will work as before.

Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>
Change-Id: I7db64603d785b2059536e371db397a11c7f0dc86