linux/ena: Fix Ubuntu compilation due to missing dma_unmap_page_attrs()
Background
==========
Commit [1] replaced the older representation of dma attributes
from struct dma_attr to an unsigned long of flags.

Commit [2] added dma_unmap_page_attr(), which receives dma attributes
as unsigned long.

Both of these commits are needed for the code to compile
without any changes.

Ubuntu issues and solutions
===========================
There are 3 possibilities:
1. Both commits [1] and [2] are applied: dma_unmap_page_attrs() exists
   and it also uses unsigned long attrs.
   In this case no kcompat changes are necessary.
2. commit [1] is applied but commit [2] isn't: dma_unmap_page_attrs()
   does not exist but unsigned long attrs are used in other functions.
   In this case we need to define dma_unmap_page_attrs() which receives
   unsigned long attrs.
3. Both commits [1] and [2] are not applied: dma_unmap_page_attrs() does
   not exist and struct dma_attrs is still used to represent dma
   attributes.
   In this case we need to define dma_unmap_page_attrs() which receives
   struct dma_attrs attrs.

Looked at ubuntu git repositories for versions of ubuntu that have amis
in EC2 ([3],[4]), to find out in which kernel version ranges fit which
of the above 3 cases.

The findings are as follows:
---------------------------
______________________________________________________________________________
| Ubuntu version | Kernel version | unsigned long attr | dma_unmap_page_attr |
==============================================================================
| 18.04 Bionic   | 4.14.0-1001    | yes                | yes                 |
------------------------------------------------------------------------------
| 16.04 Xenial   | 4.4.0-1128     | no                 | no                  |
------------------------------------------------------------------------------

Simplification of conditions
----------------------------
Since kernels in every ubuntu version start with the same X.Y.0 kernel
version, we use conditions that include future releases as well.
For example, we don't assume that 4.4.0-1128 is the last 4.4.0 kernel.
Instead we compare with kernels that are smaller than 4.5.0 to take
all 4.4.0 kernels.

Solution conclusion
-------------------
From the above table and simplifications it follows that:

1. kernels >= 4.14.0 will work without kcompat changes.
2. kernels < 4.5.0 need a definition of dma_unmap_page_attrs()
   which receives struct dma_attrs attrs.

Note 1
``````
Conclusion 1 does not add anything to the code, since we do not
support ubuntu kernels between 4.5 and 4.14, and therefore all
kernels that are >= 4.5 (which we support) will work without any
changes.

Note 2
``````
We use the definition of ENA_DMA_ATTR_SKIP_CPU_SYNC for this Ubuntu
solution similarly as it is used for RHEL.

This commit incorporates all the above mentioned changes to fix
Ubuntu compilation issues.

[1] - https://github.com/torvalds/linux/commit/00085f1efa387a8ce100e3734920f7639c80caa3
[2] - https://github.com/torvalds/linux/commit/0495c3d367944e4af053983ff3cdf256b567b053
[3] - https://code.launchpad.net/~canonical-kernel/ubuntu/+source/linux-aws/+git/xenial
[4] - https://code.launchpad.net/~canonical-kernel/ubuntu/+source/linux-aws/+git/bionic

Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>
Change-Id: I216faf0dac92937591a03856c02c98e280f5646f