fbsd/ena: Remove usage of IFF_KNOWSEPOCH
This commit is a backport of upstream commit [1].

Epoch is a concurrency mechanism used in FreeBSD.
For more details see [2].
FreeBSD network drivers have been moving to use net_epoch for handling
code that calls the network stack.
The only change needed to be done in the ena driver was to move from
TASK_INIT() to NET_TASK_INIT(). This was done in ena commit [3],
which is a backport of upstream commit [4].
Commit [5] marked the driver to be epoch ready by marking it with
IFF_KNOWSEPOCH as was instructed in upstream commit [6].
Upstream commit [1] finally removes the mark of IFF_KNOWSEPOCH, from all
drivers that moved to epoch, and uses a new flag, IFF_NEEDSEPOCH, to
mark drivers that are not epoch ready, and for those drivers upon first
call to the network stack, an error message is printed to dmesg.

Since the ena driver is epoch ready, this commit only removes the use
of IFF_KNOWSEPOCH starting from the relevant kernel version.

[1]: upstream commit a6b55ee6be15 ("net: replace IFF_KNOWSEPOCH with IFF_NEEDSEPOCH")
[2]: https://wiki.freebsd.org/Epoch
[3]: github commit 54ec0e25fa98 ("fbsd/ena: update ENA FreeBSD driver to v2.1.1")
[4]: upstream commit 6c3e93cb5a4a ("Use NET_TASK_INIT() and NET_GROUPTASK_INIT() for drivers that process incoming packets in taskqueue context.")
[5]: github commit 79095936ec27 ("fbsd/ena: update ENA FreeBSD driver to v2.2.0")
[6]: upstream commit e87c4940156c ("switch over to opt-in instead of opt-out for epoch")

Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>