linux/ena: Backport devl_ functions for devlink
Commit [1] introduced a new family of devlink param functions starting
with 'devl_' instead of 'devlink_'. The former notation implies a
lockless operation (or an operation for which a lock has been taken
before) while the latter includes taking devlink lock before doing the
operation.

The set/get commands for devlink params don't require a lock (and don't
have a devlink_ prefixed function) and so this patch modifies all
instances of
    devlink_param_driverinit_value_get
    devlink_param_driverinit_value_set

into
    devl_param_driverinit_value_get
    devl_param_driverinit_value_set

[1]: https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net-next.git/commit/?id=075935f0ae0f

Signed-off-by: Shay Agroskin <shayagr@amazon.com>