linux/ena: Fix compilation issues in RHEL 8.8
RHEL 8.8 was recently published (See [1]).
This patch fixes a compilation issue with netif_napi_add,
which was fixed for the upstream kernel in commit [2].

As RHEL 8.9 (and further) may be released in the future,
and the fact that RHEL 9.0-9.1 don't contain this backport,
the RHEL version comparison logic in kcompat was split between
RHEL 8.x and RHEL 9.x.

[1]: https://access.redhat.com/articles/3078
[2]: commit 587c738b7dac ("linux/ena: drop the weight argument from netif_napi_add")

Signed-off-by: David Arinzon <darinzon@amazon.com>