linux/ena: Fix different measurement units in comparison
When checking if the time since last napi occurred is longer than the TX
completion timeout, there is a comparison between usecs and msecs, which
is wrong, and can lead to reset due to false suspect of poll starvation.

This patch fixes the comparison and uses jiffies as the compared units.

Signed-off-by: Shahar Itzko <itzko@amazon.com>