linux/ena: Add msecs since last interrupt info for debug
Added last_intr_jiffies field in struct ena_napi which holds jiffies
value from the last occurrence of an interrupt (when last interrupt
handler was triggered), or 0 if no interrupt has ever arrived.

Removed first_interrupt field from struct ena_napi, since this
information can be provided by checking last_intr_jiffies value: If the
value of last_intr_jiffies is 0, it indicates that there was no first
interrupt. Conversely, if the value is non-zero, it implies that first
interrupt was received.

The READ_ONCE macro is employed to ensure that values are re-read during
runtime in the loop even if the compiler statically analyzed them and
determined that they are not altered within the loop.

Also updated a message to include msecs since last napi.

Signed-off-by: Shahar Itzko <itzko@amazon.com>