linux/ena: Add ethtool support for modifying large LLQ
This patch adds the capability to modify the maximum
number of bytes of a TX packet's payload the
driver writes directly to the underlying device.

The new ethtool parameter is called tx-push-buffer-len and in ENA it
can have the values 96 and 224 where the former indicates regular
entry size LLQ and the latter large LLQ header.

The upstream patchset which adds the parameter support to the kernel
can be found in [1].

The 40 magic number when formatting a string was chosen as a very
wide safe margin to the length of the string ", 224"
(which is one of the support values for tx_push_buf reported to
the user).

This commit also adds macros which define
LLQ entry's header part size in bytes.

The header part size is calculated by
    (LLQ_ENTRY_SIZE - DESC_NUMBER * DESC_LEN)

While DESC_LEN is constant (size of a struct) the other two variables
are received from the device and can theoretically change. For example
- LLQ entry size can be either 128 or 256
- number of descriptors can vary from 1 to 8, but currently
  it's set to 2

[1]:
https://lore.kernel.org/netdev/20230320132523.3203254-1-shayagr@amazon.com/

Signed-off-by: Shay Agroskin <shayagr@amazon.com>