linux/ena: Advertise TX push support on ethtool
TX push is a boolean value indicating whether the driver is writing
TX descriptors or/and their payload to the NIC
(as oppose to host's memory).

In ENA case the support is determined by whether the card supports LLQ
or not. While LLQ support can be toggled, this option is not exposed
to userspace and the driver rejects an attempt to change it.

TX push support can be queried using ethtool:
   ethtool -g [interface]

Signed-off-by: Shay Agroskin <shayagr@amazon.com>