linux/ena: Improve XDP_TX throughput by using batching
Current code issues a doorbell for every XDP_TX packet added
to the xdp(_tx) queue.
This is wasteful in case there is a burst of XDP_TX packets that
needs handling.

This commit makes sure that a doorbell is issued only at the end
of a napi call.

This commit also removes the flags parameter to ena_xdp_xmit_frame()
and the code that uses it in the function, as of the time
this commit was introduced only flags=0 are passed to the
function, making it useless.

Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>