linux/ena: Set tx_info->xdpf value to NULL
Commit [1] removed the explicit assignment of tx_info->xdpf to NULL
with the following justification:

    No need to set tx_info->xdpf to NULL and tx_info->num_of_bufs to 0
    in case of a mapping error. Both values won't be used once the
    mapping function returns an error, and their values would be
    overridden by the next transmitted packet.

While both values do indeed get overridden in the next transmission
call, the value of tx_info->xdpf is also used to check whether a TX
descriptor's transmission has been completed (i.e. a completion for it
was polled).

An example scenario:
1. Mapping failed, tx_info->xdpf wasn't set to NULL
2. A VF reset occurred leading to IO resource destruction and
   a call to ena_free_tx_bufs() function
3. Although the descriptor whose mapping failed was freed by the
   transmission function, it still passes the check
     if (!tx_info->skb)

   (skb and xdp_frame are in a union)
4. The xdp_frame associated with the descriptor is freed twice

This patch returns the assignment of NULL to tx_info->xdpf to make the
cleaning function know that the descriptor is already freed.

[1]: See 'Fixes:' tag

Fixes: 2ddf43dbb96a ("linux/ena: Fix mapping function issues in XDP")
Signed-off-by: Shay Agroskin <shayagr@amazon.com>