linux/ena: Reduce print level of strscpy truncation warnings
With the introduction of strscpy usage in commits [1] and [2], when the
source string is longer than the destination string, the API omits
a truncation warning.
Given that this warning doesn't have any functional impact,
meaning, a user doesn't need to make any change to overcome it,
and it's merely informative, the print level of this
warning is reduced to debug in order to avoid unnecessary concerns
due to logging.

[1]: commit 0320d090d6c3 ("linux/ena: Move from strlcpy with unused retval to strscpy")
[2]: commit 12e8e6a8a482 ("linux/ena: Add status check for strscpy calls")

Signed-off-by: David Arinzon <darinzon@amazon.com>