linux/ena: Check that recorded RX queue is within range
The ena_select_queue() function is called by the kernel to choose what
TX queue should transmit a given SKB. The function returns a queue id.

The function chooses the TX queue in one of two ways:
    - If the packet that needs to be transmitted has a record of the RX
      queue on which it was received (e.g. if the packet was received on
      RX queue 1) then the driver would transmit the packet from the
      same queue
    - If such record doesn't exist, the driver (using a kernel helper
      function) uses the packet's hash to choose the queue.

If the driver is the one recording the RX queue then
using it as TX queue id is okay (we never have more RX queues than TX),
however, if another entity records it then the recorded
RX queue might not be in the same range as the TX queues.

In case the driver returns a TX queue id that isn't within the TX
queues range the kernel prints a warning and transmits the packet
through queue 0.

This patch makes the driver check the recorded RX queue before choosing
it as TX.

Signed-off-by: Shay Agroskin <shayagr@amazon.com>