linux/ena: Extend support to 256 reset reasons
Currently we only use 4 bits in order to define the reset reason.
This allows us to define 16 reset reasons which have already been used
up. In order to allow for more reset reasons to be defined and detected,
this patch will extend the reset reason up to 8 bits which allows
256 reset reasons.

For backward compatibility reasons, bits 28-31 will represent the LSB
and bits 24-27 will represent the MSB in the relevant register.
This way when an old driver writes a reset reason it will be compatible
with the newer devices.

Also, in case the device does not support extended reset reasons, the
driver will identify it and fall back to GENERIC reset reason.
When using an old driver all reset reasons are mapped correctly
except of ENA_REGS_RESET_RX_DESCRIPTOR_MALFORMED since it was
already defined with index 16 before extending the reset
reasons support, in this case the device will interpret the
reset reason as ENA_REGS_RESET_NORMAL.

Signed-off-by: Osama Abboud <osamaabb@amazon.com>