linux/ena: Add configuration notifications interface support
This commit is part of the effort of notifying the user of non-optimal
or performance impacting practices.

A new interface was introduced to serve as a communication channel
between the device and the driver. One of the goals of this channel is
to create a new mechanism for notifying the driver and user in case of
sub-optimal configuration.

This commit adds the needed defines in ena_admin_defs and the needed
handler in order to support this channel on the driver side.

Signed-off-by: Osama Abboud <osamaabb@amazon.com>