linux/ena: Backport exponential backoff exp limit
This patch backports a change made by the community in commit
[1], which limits the exponent in the exponential backoff
mechanism in order to avoid the value overflowing.
The overflow causes a panic on the customer's system.

For uniformity and readability purposes, the min/max parameter
in the calls of ENA_MIN32 and ENA_MAX32 macros was changed
to be first.

[1]: https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net-next.git/commit/?id=1e9cb763e9bacf0c932aa948f50dcfca6f519a26

Signed-off-by: David Arinzon <darinzon@amazon.com>