linux/ena: Fix compilation issue in Fedora 37 and 38
In recent releases of Fedora 37 and 38, RHEL_RELEASE_CODE
was added to the generated headers (See [1]), as Fedora
is the upstream version of RHEL (See [2]).

To be able to distinguish itself from RHEL, the versions
were defined as RHEL_MAJOR=9, RHEL_MINOR=99 (See [3]).
This causes a compilation issue where RHEL_RELEASE_CODE
has been checked.

As Fedora closely follows the upstream kernel releases
(See [4]), it is possible to rely on LINUX_VERSION_CODE
macro for version comparison.

This change is fixing a compilation issue identified
in [5] by adding an explicit check for FEDORA_RELEASE
to identify Fedora as well as a more accurate check
for the devlink feature advertisement removal, which
was introduced in kernel 6.3 (See [6]).

[1]: https://gitlab.com/cki-project/kernel-ark/-/commit/37aa3c0686feb58dc8fc3ed1722acc28c95d5d7a
[2]: https://docs.fedoraproject.org/en-US/quick-docs/fedora-and-red-hat-enterprise-linux/
[3]: https://gitlab.com/cki-project/kernel-ark/-/commit/c0409891ace5528d4b786b78c09b8e4041210a72
[4]: https://docs.fedoraproject.org/en-US/quick-docs/kernel/overview/
[5]: https://github.com/amzn/amzn-drivers/issues/273
[6]: https://github.com/torvalds/linux/commit/fb8421a94c5613fee86e192bab0892ecb1d56e4c

Signed-off-by: David Arinzon <darinzon@amazon.com>