linux/ena: Initialize llq_policy field for all cases
Commit [1] added the field llq_policy to struct ena_adapter,
which is being accessed in the code, and initialized to 0,
but the driver relies on one of the enum values to be 0,
which is not granted.

In case there is no such value in the enum, the field will be
initialized correctly.
There are two cases:
- force_large_llq_header is set: initialization in [2]
- force_large_llq_header is not set: wrong initialization to 0

This commit adds an explicit initialization to the field
for every case.

[1]: See Fixes
[2]: ena_set_forced_llq_size_policy() in ena_netdev.c

Fixes: 5366abefe205 ("linux/ena: Support LLQ entry size recommendation from device")
Signed-off-by: Shahar Itzko <itzko@amazon.com>