linux/ena: Fix skb->truesize in case the RX buffer is not reused
skb->truesize is used for socket memory usage bookkeeping.
skb->truesize should represent the amount of memory used by the RX
buffer even if not all of it is used by the packet itself. (see [1]
for more details)

If the driver doesn't set skb->truesize with the correct amount of
memory taken by the skb, it may cause memory usage issues.

Before introducing DRB, the amount of memory used by each buffer
in an RX packet was ENA_PAGE_SIZE, which is the size of the whole
RX buffer, even if only a small part of the buffer was actually used
by the packet. To make sure skb->truesize reflected this, ENA_PAGE_SIZE
was passed as the frag_size to build_skb().
Then build_skb()->..->__finalize_skb_around() calculated the correct
skb->truesize for this skb.

Since DRB was introduced in [2] skb->truesize is no longer calculated
correctly for buffers that are not reused.

When DRB recycles a buffer, buf_len is the size of the part of the
buffer used by the packet, but it is also the only part of the buffer
used for this packet (the rest is recycled for the next packet).
Therefore, in this case buf_len is the correct value for the frag size
parameter passed to build_skb().

But when DRB doesn't recycle a buffer, buf_len is (usually) not the
correct value to be passed to build_skb().

For example if we have a 4K page RX buffer, and buf_len is 3K, DRB
will not recycle this page (since there isn't enough room left in the
buffer). In such a case, the frag_size parameter to build_skb should be
4k because all of the 4k are being used by this packet. However, the
current implementation incorrectly passes buf_len=3K to build_skb.

This patch fixes the buf_len passed to build_skb() for the first buffer
of the packet and to skb_add_rx_frag() for the following buffers in case
DRB doesn't recycle the RX buffer.
It passes the size of the whole part of the RX buffer that is left after
previous recycles of this buffer (if any).

[1]: http://vger.kernel.org/~davem/skb_sk.html
[4]: commit 7d7dbf0b4a9a ("linux/ena: Add Dynamic RX Buffer feature (DRB)")

Fixes: 1dd2cba919a6 ("linux/ena: Fix RHEL7 compilation issues due to missing dma_unmap_page_attrs()")
Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>