linux/ena: Fix large LLQ in ethtool when devlink is disabled
Due to a bad rebase on top of large LLQ recommendation patch [1] the
value of 'adapter->llq_policy' wasn't changed to reflect changes in LLQ
entry size when modified using ethtool [2]. This can result in leaving
LLQ entry size the same.

The bug is mitigated in case devlink is configured
(i.e. has CONFIG_DEVLINK set in its kernel and the kernel version is
newer than 4.19).

[1]: commit 5366abefe205("linux/ena: Support LLQ entry size recommendation from device")
[2]: See Fixes

Fixes: e89ffd531e90 ("linux/ena: Add ethtool support for modifying large LLQ")
Signed-off-by: Shay Agroskin <shayagr@amazon.com>