fbsd/ena: Introduce __FreeBSD_version compilation flag
Currently, there is __FreeBSD_version usage in multiple files, some of
which are wrapping the licensing headers.
This is causing some issues since this define is defined in
sys/param.h header file in the kernel and we wish to use it before we're
including the sys/param.h header in the file.

This commit adds this define as a compilation flag through using sysctl
command and creating the relevant flag.

Signed-off-by: Osama Abboud <osamaabb@amazon.com>