linux/ena: Fix rpm installation for RHEL 8.x and 9.x
RHEL supports updating the driver via rpm.
Changes were made in RHEL 8.x and 9.x and as a result
users couldn't install the driver for those versions via rpm.

This commit adds necessary code and updates the README-rpm in order
to support installing the driver via rpm on the new RHEL versions.

Signed-off-by: Ofir Tabachnik <ofirt@amazon.com>