linux/ena: Support max wide LLQ depth from device
Wide LLQ depth size is currently calculated by dividing the maximum
possible size of LLQ by 2.
From 7th gen, on selected instances, the size of BAR2,
which contains LLQ, will be increased, and the maximum depth of wide
LLQ will be set according to a value set by the device, instead of
hardcoded division by 2.

The new value will be stored by the device in max_wide_llq_depth field
for drivers that expose ENA_ADMIN_LLQ_FEATURE_VERSION_1 or
higher to the device.

There is an assumption that max_llq_depth >= max_wide_llq_depth, since
they both use the same bar, and if it is possible to have a wide LLQ
of size max_wide_llq_depth, it is possible to have a normal LLQ of the
same size, since it will occupy half of the space.

The wide LLQ's default size will remain 512 instead of utilizing
the maximum size.

Also moved the large LLQ case calculation of max_tx_queue_size
before its rounddown and the check for minimal ring size.

Signed-off-by: Shahar Itzko <itzko@amazon.com>