linux/ena: Document large LLQ enablement by default
Starting with 6th generation, large LLQ is enabled by
default due to the device recommendation.
As customers might be facing issues (See [1]) in which the
Tx ring size is decreased to 512 without the ability to
increase back to 1024 without disabling large LLQ first,
additional documentation is needed to notify about this
change and add information on how to change back to
standard LLQ with 1024 Tx ring size.
Also adding the documentation about the fact that Tx ring
size may be increased back to 1024 on 7th generation
instance-types.
Large LLQ configuration via ethtool was added in upstream kernel
v6.4. (See tags in [1]).
This commit also fixing the documentation that says that the minimal
required kernel version is v6.3.

[1]: https://github.com/torvalds/linux/commit/b0c59e53968b55f6add088319bd31ca7dcdd8484

Signed-off-by: David Arinzon <darinzon@amazon.com>