linux/ena: Verify number of descriptors for copybreak
Copybreak logic assumes that a packet fits into one descriptor.
Copybreak size can be changed by the user, and if the length of the
first descriptor is less than or equal to copybreak value, the driver
will choose the copybreak path, even though the packets could be spread
across multiple descriptors. This results in dropped packets.
This commit adds a verification that the number of descriptors is one
as a condition to enter copybreak logic.

Signed-off-by: Ofir Tabachnik <ofirt@amazon.com>