linux/ena: Handle ENA_CMD_ABORTED case on admin queue interrupt mode
Currently admin_queue->stats.aborted_cmd counter is incremented if an
admin command status is ENA_CMD_ABORTED and only if the admin queue is
in polling mode.
This commit fixes handling the case of incrementing
admin_queue->stats.aborted_cmd if the admin queue is in interrupt
mode as well.

Signed-off-by: Ofir Tabachnik <ofirt@amazon.com>