linux/ena: Fix possible stuck tx packets when a tx drop occurs
This fix was suggested in [1].

In case there is a tx packet burst, where N packets are to be sent,
if the last packet in the burst is dropped in ena_start_xmit(),
the doorbell is not sent to the device for this burst.

In such a case, the first N-1 packets in the burst will not be sent
until the next time a doorbell is sent.

This commit fixes this issue by sending a doorbell after a dropped
packet if it is the last in a tx burst !netdev_xmit_more() and also
after checking that indeed there are descriptors sent to the device.

[1]: https://lore.kernel.org/netdev/CANn89iJRHuFgPrsezhm2DuAw7JmLL2ZkPhZaf2Ymuq+STUm-8w@mail.gmail.com/

Suggested-by: Eric Dumazet <edumazet@google.com>
Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>