linux/ena: Modify Numa node update on io_poll
The device supports a PCIe optimization hint, which indicates on
which NUMA the queue is currently processed. This hint is utilized
by PCIe (as long as the cpu host supports this feature as well)
in order to reduce its access time by accessing the correct NUMA
resources and maintaining cache coherence, when copying rx buffers to
the host. This functionality is applicable for rx rings only, also this
is XDP indenepdent.

This commit updates the following:
1. Remove redundant tx_ring->numa_node update.
2. Update rx_ring for AF_XDP as well.

This commit also wraps the xsk ring check in #ifdef ENA_AF_XDP_SUPPORT,
since we don't want this code running if AF_XDP is not supported.

Signed-off-by: Evgeny Ostrovsky <evostrov@amazon.com>
Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>